requirements.txt
django==5.2.11
djangorestframework==3.16.1
mysqlclient==2.2.7
python-decouple==3.8
openpyxl==3.1.5
pytest==9.0.2
pytest-django==4.11.1
coverage==7.13.3
factory-boy==3.3.3
Faker==40.1.2